Organize tests and fix Jet test
